LayoutTests/imported/w3c/web-platform-tests/css/css-ui/parsing/canonical-order-outline-sub-properties-001-expected.txt


PASS testing outline: blue
FAIL testing outline: invert assert_equals: expected "3px none invert" but got "3px none rgb(0, 0, 255)"
FAIL testing outline: 4px assert_equals: expected "4px none invert" but got "4px none rgb(0, 0, 0)"
FAIL testing outline: solid assert_equals: expected "3px solid invert" but got "3px solid rgb(0, 0, 0)"
FAIL testing outline: solid 5px assert_equals: expected "5px solid invert" but got "5px solid rgb(0, 0, 0)"
FAIL testing outline: 6px dashed assert_equals: expected "6px dashed invert" but got "6px dashed rgb(0, 0, 0)"
PASS testing outline: dotted blue
FAIL testing outline: dotted invert assert_equals: expected "3px dotted invert" but got "3px dotted rgb(0, 0, 255)"
PASS testing outline: blue solid
FAIL testing outline: invert solid assert_equals: expected "3px solid invert" but got "3px solid rgb(0, 0, 255)"
PASS testing outline: black 4px
FAIL testing outline: invert 4px assert_equals: expected "4px none invert" but got "4px none rgb(0, 0, 0)"
PASS testing outline: 5px blue
FAIL testing outline: 5px invert assert_equals: expected "5px none invert" but got "5px none rgb(0, 0, 255)"
PASS testing outline: black solid 6px
FAIL testing outline: invert solid 6px assert_equals: expected "6px solid invert" but got "6px solid rgb(0, 0, 0)"
PASS testing outline: blue 4px dotted
FAIL testing outline: invert 4px dotted assert_equals: expected "4px dotted invert" but got "4px dotted rgb(0, 0, 255)"
PASS testing outline: dashed 5px black
FAIL testing outline: dashed 5px invert assert_equals: expected "5px dashed invert" but got "5px dashed rgb(0, 0, 0)"
PASS testing outline: solid blue 6px
FAIL testing outline: solid invert 6px assert_equals: expected "6px solid invert" but got "6px solid rgb(0, 0, 255)"
PASS testing outline: 4px dotted black
FAIL testing outline: 4px dotted invert assert_equals: expected "4px dotted invert" but got "4px dotted rgb(0, 0, 0)"
PASS testing outline: 5px blue dashed
FAIL testing outline: 5px invert dashed assert_equals: expected "5px dashed invert" but got "5px dashed rgb(0, 0, 255)"
